// src/api/generated/model/messageCollectionItemDto.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Ludora API
 * API for the Ludora app
 * OpenAPI spec version: 0.0.1
 */
import type { MessageCollectionItemDtoGlobalStatus } from './messageCollectionItemDtoGlobalStatus';
import type { MessageCollectionItemDtoType } from './messageCollectionItemDtoType';

export interface MessageCollectionItemDto {
  /** Message content */
  readonly content: string;
  /** Message creation date */
  readonly createdAt: string;
  /** Message global status */
  readonly globalStatus: MessageCollectionItemDtoGlobalStatus;
  /** Whether the message has been read by anyone */
  readonly hasAnyRead: boolean;
  /** Whether the message has been read by everyone */
  readonly hasEveryoneRead: boolean;
  /** Message type */
  readonly type: MessageCollectionItemDtoType;
  /** Message ID */
  readonly uid: string;
}