// src/api/generated/model/paginationResponseMessageCollectionItemDto.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Ludora API
 * API for the Ludora app
 * OpenAPI spec version: 0.0.1
 */
import type { PaginationDataMessageCollectionItemDto } from './paginationDataMessageCollectionItemDto';

export interface PaginationResponseMessageCollectionItemDto {
  readonly data: PaginationDataMessageCollectionItemDto;
  readonly message?: string;
}